requirements.txt
fastapi
uvicorn
python-multipart
jinja2
google-genai
opencv-python-headless
numpy
python-dotenv
pillow
